requirements.txt
streamlit==1.52.2
pandas==2.3.3
numpy==2.4.1
openpyxl==3.1.5
rapidfuzz==3.14.3
xlsxwriter==3.2.9
python-dateutil==2.9.0.post0
